fix: extract form utils to fix ESLint react-refresh rule and add ESLint to CI
Move non-component exports (form state types, converters, validators) out
of MonitorForm.tsx and StepForm.tsx into dedicated utility modules so that
react-refresh/only-export-components is satisfied. Add ESLint step to the
frontend CI job.
