+++ .github/workflows/test.yaml
@@ -33,5 +33,6 @@ jobs:
           cache: npm
           cache-dependency-path: frontend/package-lock.json
       - run: npm ci
+      - run: npx eslint src
       - run: npm run build
       - run: npm test
