+++ frontend/src/__tests__/MonitorForm.test.tsx
@@ -3,12 +3,13 @@ import React from 'react'
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import MonitorForm, {
+import MonitorForm from '../components/MonitorForm'
+import {
   formStateToMonitor,
   monitorToFormState,
   validateForm,
   emptyFormState,
-} from '../components/MonitorForm'
+} from '../components/monitor-form-utils'
 
 vi.mock('../api')
